requirements.txt
fuzzywuzzy==0.17.0
gensim==3.5.0
nltk==3.3
numpy==1.14.5
PyYAML==5.1
torch==0.4.1

